Efl: make eeze optional. Need the bug URL (OOPS!).
Eeze is the EFL wrapper on top of udev and is only used in two cases:
enumerate gamepads and check for 'raw' network connectivity (instead
of reaching NetworkManager or ConnMan, it will just check netlink for
a network interface with configured IP address (which doesn't mean it
can reach the internet at all). It's good enough for some cases, but
not the best solution and definitely not mandatory.

Reviewed by NOBODY (OOPS!).

:
* Source/cmake/OptionsEfl.cmake: Eeze is now optional.

Source/WebCore:

No new test are required.

* PlatformEfl.cmake: EEZE_LIBRARIES and EEZE_INCLUDE_DIRS conditional to USE(EEZE)
* platform/network/NetworkStateNotifier.h: PLATFORM(EFL) && USE(EEZE)
* platform/network/efl/NetworkStateNotifierEfl.cpp: #if USE(EEZE)